Moved print_output() to new file parser_lib.py and refactored how the individual parsers are imported
