+++ tools/wiktionary2lex/parsers/en.py
@@ -16,11 +16,13 @@
 # See the Apache 2 License for the specific language governing permissions and
 # limitations under the License.
 
-from lxml import etree as ET
 import sys
 import time
 import re
 import pickle
+from lxml import etree as ET
+
+import parser_lib
 
 def extract_entries(wiktionary_xml):
     if __debug__:
@@ -159,15 +161,6 @@ def add_to_output(output,word,pronunciation,POS,language,accent,xsampa):
                 "POS":None, "lang":language, 
                 "accent":accent, "x-sampa":xsampa})
 
-def print_output(output):
-    for list_item in output:
-        for key in list_item:
-            if list_item[key] is None:
-                print key + ":None\t",
-            else:
-                print key + ":" + list_item[key] + "\t",
-        print "\n"
-
 if __name__ == "__main__":
     reload(sys)
     sys.setdefaultencoding('utf-8')
@@ -181,4 +174,4 @@ if __name__ == "__main__":
         output_filename = "output/output_" + timestr + ".txt"
         output_file = open(output_filename,'wb')
         pickle.dump(output,output_file)
-        print_output(output)
+        parser_lib.print_output(output)
