+++ tools/wiktionary2lex/wiktionary2lex.py
@@ -1,20 +1,38 @@
 #!/usr/bin/env python2
 # -*- coding: utf-8 -*-
 
-from third_party import cxs
+# Copyright 2018 Cereproc Ltd. (author: Caoimhín Laoide-Kemp)
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#  http://www.apache.org/licenses/LICENSE-2.0
+#
+# THIS CODE IS PROVIDED *AS IS* BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, EITHER EXPRESS OR IMPLIED, INCLUDING WITHOUT LIMITATION ANY IMPLIED
+# WARRANTIES OR CONDITIONS OF TITLE, FITNESS FOR A PARTICULAR PURPOSE,
+# MERCHANTABLITY OR NON-INFRINGEMENT.
+# See the Apache 2 License for the specific language governing permissions and
+# limitations under the License.
+
 import imp
 import sys
-from lxml import etree as ET
 import itertools
 import argparse
+from lxml import etree as ET
+
+from third_party import cxs
+
+from parsers import parser_lib
 
 def ipa2xsampa(ipa_text,table):
     xsampa_text = cxs.ucipa2cxs(ipa_text,table)
     return xsampa_text
 
-def xml2dict(language_code,xml):
-    parser_module = imp.load_source("parser","parsers/" + language_code + ".py")
-    output = parser_module.extract_entries(xml)
+def xml2dict(language_code,wiktionary_xml):
+    extract_entries = parser_lib.get_parser(language_code)
+    output = extract_entries(wiktionary_xml)
     return output
 
 def add_xsampa_to_data(parsed_data):
@@ -23,15 +41,6 @@ def add_xsampa_to_data(parsed_data):
     for entry in parsed_data:
         if entry["pron"] is not None:
             entry["x-sampa"] = ipa2xsampa(entry["pron"],table)
-
-def print_output(output):
-    for list_item in output:
-        for key in list_item:
-            if list_item[key] is None:
-                print key + ":None\t",
-            else:
-                print key + ":" + list_item[key] + "\t",
-        print "\n"
 
 def output2lex(output,filename):
     lexicon_output = open(filename,"wb")
@@ -93,6 +102,6 @@ if __name__ == "__main__":
 
     output = xml2dict(language_code,wiktionary_xml)
     add_xsampa_to_data(output)
-    print_output(output)
+    parser_lib.print_output(output)
     filtered_output=filter_list(output,"x-sampa","word")
     output2lex(filtered_output,output_lexicon)
